requirements.txt
ffmpeg-python
audio-recorder-streamlit
gtts
pydub
pyttsx3
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.26
langchain-groq==0.1.9
python-dotenv==1.0.1
streamlit==1.37.0
langchain_text_splitters
langchain_huggingface
faiss-cpu
pypdf